♻️ Refactor: 파일 구조 수정 및 ui 개선
