--- src/components/CrewSchedule/button/post.tsx
@@ -1,9 +1,0 @@
-const Post = () => (
-    <button
-        className="bg-[#3A3ADB] hover:bg-blue-700 text-white text-sm font-bold py-1 px-8 rounded"
-      >
-        글쓰기
-      </button>
-);
- 
-export default Post;
