+++ src/components/detail/Schedule/button/post.tsx
@@ -1,0 +1,9 @@
+const Post = () => (
+    <button
+        className="bg-[#3A3ADB] hover:bg-blue-700 text-white text-sm font-bold py-1 px-8 rounded-sm"
+      >
+        글쓰기
+      </button>
+);
+ 
+export default Post;
