+++ src/pages/detail/index.tsx
@@ -5,19 +5,15 @@ import AboutSection from "../../components/detail/about";
 
 function Detail() {
   return (
-    <div className = "bg-gray-100">
-      <div className="py-6 space-y-6 pt-12"> 
-      <div>
-      <Header />
-      <Tabs />
+    <div className="bg-gray-100 min-h-screen">
+      <div className="mt-12">
+        <Header />
+        <Tabs />
       </div>
-      <div className="px-6 py-6 space-y-3 pt-0"> 
-      <Notice />
-      <div className = "pt-3">
-      <AboutSection />
+      <div className="w-full max-w-screen-xl mx-auto px-4 sm:px-6 lg:px-12 xl:px-20 pt-6 space-y-6">
+        <Notice />
+        <AboutSection />
       </div>
-      </div>
-    </div>
     </div>
   );
 }
